sifter: apply 2to3 changes
With Python 2.7 going EOL soon, conversion to Python 3 is necessary.

+++ sifter/validators/number.py
@@ -8,7 +8,7 @@ class Number(sifter.grammar.Validator):
         if starting_index >= len(arg_list):
             return 0
         try:
-            long(arg_list[starting_index])
+            int(arg_list[starting_index])
             return 1
         except TypeError:
             return 0
